services/mfp.py
"""MyFitnessPal API client."""
import aiohttp
from typing import Optional, Dict
from datetime import date
import logging

logger = logging.getLogger(__name__)


class MyFitnessPalClient:
    """
    Client for MyFitnessPal integration.

    Note: MyFitnessPal's public API is limited. This is a stub implementation.
    You may need to use:
    1. MyFitnessPal Premium API (paid)
    2. Third-party libraries like `myfitnesspal` Python package
    3. Web scraping (not recommended)
    4. Apple Health / Google Fit as intermediary
    """

    BASE_URL = "https://api.myfitnesspal.com/v2"

    def __init__(self, username: str = "", password: str = ""):
        self.username = username
        self.password = password
        self._session = None

    async def authenticate(self) -> bool:
        """Authenticate with MyFitnessPal."""
        # TODO: Implement actual authentication
        # This will depend on which method you choose (API, scraping, etc.)
        logger.warning("MyFitnessPal authentication not yet implemented")
        return False

    async def get_nutrition(
        self,
        target_date: Optional[date] = None
    ) -> Optional[Dict]:
        """
        Get nutrition data for a specific date.

        Args:
            target_date: Date to fetch nutrition for (defaults to today)

        Returns:
            Nutrition data dictionary or None
        """
        if not target_date:
            target_date = date.today()

        # TODO: Implement actual data fetching
        # Options:
        # 1. Use myfitnesspal Python package
        # 2. Implement API calls (if you have premium access)
        # 3. Parse from web scraping

        logger.warning(f"MyFitnessPal data fetch not yet implemented for {target_date}")

        # Return mock data structure for now
        return {
            "date": target_date.isoformat(),
            "calories": 0,
            "protein": 0,
            "carbs": 0,
            "fat": 0,
            "fiber": 0,
            "source": "myfitnesspal"
        }

    async def get_weight(self, target_date: Optional[date] = None) -> Optional[float]:
        """Get weight measurement for a date."""
        # TODO: Implement weight tracking
        return None

    def parse_nutrition(self, raw_data: Dict) -> Dict:
        """
        Parse MyFitnessPal data into our format.

        Args:
            raw_data: Raw nutrition data

        Returns:
            Parsed nutrition dictionary
        """
        return {
            "date": raw_data.get("date"),
            "calories": raw_data.get("calories", 0),
            "protein": raw_data.get("protein", 0),
            "carbs": raw_data.get("carbohydrates", 0),
            "fat": raw_data.get("fat", 0),
            "fiber": raw_data.get("fiber", 0),
            "source": "myfitnesspal"
        }


# Alternative using the myfitnesspal package
# Uncomment and add 'myfitnesspal' to requirements.txt to use
"""
import myfitnesspal

class MyFitnessPalPackageClient:
    def __init__(self, username: str, password: str):
        self.client = myfitnesspal.Client(username, password)

    def get_nutrition(self, target_date: date = None) -> Dict:
        if not target_date:
            target_date = date.today()

        day = self.client.get_date(target_date)
        totals = day.totals

        return {
            "date": target_date.isoformat(),
            "calories": totals.get("calories", 0),
            "protein": totals.get("protein", 0),
            "carbs": totals.get("carbohydrates", 0),
            "fat": totals.get("fat", 0),
            "fiber": totals.get("fiber", 0),
            "source": "myfitnesspal"
        }
"""
