requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiohttp==3.9.1
google-generativeai==0.3.2
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
